requirements.txt
pydantic>=2.0.0
openai>=1.0.0
gradio>=4.0.0
pyyaml>=6.0
